requirements.txt
absl-py=2.1.0
chex=0.1.85
distrax=0.1.5
dm-haiku=0.0.11
dm-pix=0.4.2
immutabledict=4.1.0
jax=0.4.24
jaxlib=0.4.24
jaxline=0.0.8
ml-collections=0.1.1
numpy=1.25.2
optax=0.1.9
pillow=10.2.0
scipy=1.11.4
tensorflow=2.15.*
torch=2.2.0
torchvision=0.17.0
tqdm=4.66.1
